Move gatt classes into separate module
